requirement.txt
aiohttp==3.6.3
aiohttp-cors==0.7.0
aioredis==1.3.1
alabaster==0.7.12
alembic==1.4.3
amqp==2.6.1
aniso8601==7.0.0
ansible==2.8.3
apache-airflow==1.10.12
apispec==1.3.3
appdirs==1.4.4
argcomplete==1.12.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bleach==3.2.1
blessings==1.7
cached-property==1.5.2
cachetools==4.1.1
cattrs==1.0.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coloredlogs==14.0
colorful==0.5.4
colorlog==4.0.2
configparser==3.5.3
contextlib2==0.6.0.post1
contextvars==2.4
croniter==0.3.34
cryptography==3.0
cycler==0.10.0
dagit==0.9.11
dagster==0.9.11
dagster-graphql==0.9.11
dagster-pandas==0.9.14
dagster-pyspark==0.9.14
dagster-spark==0.9.14
dask==2.30.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
distlib==0.3.1
Django==2.1
django-rest-knox==4.1.0
djangorestframework==3.11.1
dnspython==1.16.0
docstring-parser==0.7.1
docutils==0.16
email-validator==1.1.1
entrypoints==0.3
eventlet==0.26.1
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-Cors==3.0.9
Flask-GraphQL==2.0.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SocketIO==4.3.1
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
google==3.0.0
google-api-core==1.22.4
google-auth==1.22.1
googleapis-common-protos==1.52.0
gpustat==0.6.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
graphql-ws==0.3.1
graphviz==0.14.2
greenlet==0.4.17
grpcio==1.30.0
gunicorn==20.0.4
hiredis==1.1.0
humanfriendly==8.2
hydralib==5.57.3
idna==2.5
idna-ssl==1.1.0
imagesize==1.2.0
immutables==0.14
importlib-metadata==1.7.0
ipython-genutils==0.2.0
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
jira==2.0.0
joblib==0.17.0
json-merge-patch==0.2
jsonschema==3.2.0
jupyter-core==4.6.3
kafka==1.3.5
kiwisolver==1.2.0
kombu==4.6.11
lazy-object-proxy==1.5.1
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.2
mistune==0.8.4
monotonic==1.5
msgpack==1.0.0
multidict==4.7.6
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.7
numpy==1.19.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.10
opencensus-context==0.1.1
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
pathtools==0.1.2
pbr==5.5.0
pendulum==1.4.4
Pillow==7.2.0
prison==0.1.3
prometheus-client==0.8.0
promise==2.3
protobuf==3.12.4
psutil==5.7.2
py-spy==0.3.3
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyspark==3.0.1
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.1
python-nvd3==0.15.0
python-slugify==4.0.1
python-socketio==4.6.0
python3-openid==3.2.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
ray==1.0.0
redis==3.4.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
Rx==1.6.1
scikit-learn==0.23.2
scipy==1.5.2
setproctitle==1.1.10
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.19
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
tabulate==0.8.7
tenacity==4.12.0
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
thrift==0.13.0
toolz==0.10.0
toposort==1.5
tqdm==4.50.0
traitlets==4.3.3
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.21.1
vine==1.3.0
watchdog==0.10.3
webencodings==0.5.1
Werkzeug==0.16.1
WTForms==2.3.3
yarl==1.5.1
zipp==1.2.0
zope.deprecation==4.4.0
zope.event==4.4
zope.interface==5.1.0
